Refactor to move the localization in the model project.
