+++ ResearchSuite/ResearchSuite/en.lproj/ResearchSuite.strings
@@ -36,3 +36,43 @@
 "BOOL_NO" = "No";
 "BOOL_YES" = "Yes";
 
+/* Button titles */
+"BUTTON_OK" = "OK";
+"BUTTON_CANCEL" = "Cancel";
+"BUTTON_DONE" = "Done";
+"BUTTON_GET_STARTED" = "Get Started";
+"BUTTON_LEARN_MORE" = "Learn more";
+"BUTTON_NEXT" = "Next";
+"BUTTON_BACK" = "Back";
+"BUTTON_SKIP" = "Prefer not to answer";
+"BUTTON_SKIP_TASK" = "I can't do this now";
+"BUTTON_PAUSE" = "Pause";
+"BUTTON_RESUME" = "Resume";
+
+/* Joining items */
+"TWO_ITEM_LIST_OR_FORMAT_%1$@_%2$@" = "%1$@ or %2$@"; // Format of a list with two items (For example, 'apples or oranges')
+"THREE_ITEM_LIST_OR_FORMAT_%1$@_%2$@_%3$@" = "%1$@, %2$@, or %3$@"; // Format of a list with three items (For example, 'apples, oranges, or bananas')
+"TWO_ITEM_LIST_AND_FORMAT_%1$@_%2$@" = "%1$@ and %2$@"; // Format of a list with two items (For example, 'apples and oranges')
+"THREE_ITEM_LIST_AND_FORMAT_%1$@_%2$@_%3$@" = "%1$@, %2$@, and %3$@"; // Format of a list with three items (For example, 'apples, oranges, and bananas')
+"LIST_FORMAT_DELIMITER"  = ", "; // Delimiter for a list of more than 3 items. Include whitespace as appropriate. (For example, 'foo, bar, goo, or boo')
+
+/* Progress */
+"CURRENT_STEP_%@_OF_TOTAL_STEPS_%@" = "Step %1$@ of %2$@";  // Text for marking the current steps
+
+/* Default placeholders to request answers in a questionnaire. */
+"PLACEHOLDER_TEXT_OR_NUMBER" = "Tap to answer";
+"NULL_ANSWER" = "Select an answer";
+"PLACEHOLDER_IMAGE_CHOICES" = "Tap to select";
+"PLACEHOLDER_LONG_TEXT" = "Tap to write";
+
+/* Validation Error Messages */
+"VALIDATION_ERROR_TITLE" = "Invalid Answer";
+"VALIDATION_ERROR_DATE_%@" = "Sorry, the date you entered is not in a recognized format. The expected format is %@";
+"VALIDATION_ERROR_NUMBER" = "Sorry, the number you entered is not in a recognized format.";
+"VALIDATION_ERROR_GENERIC" = "Sorry, the text you have entered is invalid."; // Generic message for invalid text entry
+"VALIDATION_ERROR_MAXLEN_%@" = "The text you entered exceeds the maximum character length of %@ characters.";
+"VALIDATION_ERROR_MIN_DATE_%@" = "The date entered is before %@.";
+"VALIDATION_ERROR_MAX_DATE_%@" = "The date entered is after %@.";
+"VALIDATION_ERROR_MIN_NUMBER_%@" = "The number entered is less than %@.";
+"VALIDATION_ERROR_MAX_NUMBER_%@" = "The number entered is more than %@.";
+
