--- ResearchSuiteUI/ResearchSuiteUI/Localization/en.lproj/ResearchSuiteUI.strings
@@ -1,92 +1,0 @@
-/* 
-  ResearchSuiteUI.strings
-  ResearchSuiteUI
-
-  Copyright © 2016 Sage Bionetworks. All rights reserved.
-
- Redistribution and use in source and binary forms, with or without modification,
- are permitted provided that the following conditions are met:
-
- 1.  Redistributions of source code must retain the above copyright notice, this
- list of conditions and the following disclaimer.
-
- 2.  Redistributions in binary form must reproduce the above copyright notice,
- this list of conditions and the following disclaimer in the documentation and/or
- other materials provided with the distribution.
-
- 3.  Neither the name of the copyright holder(s) nor the names of any contributors
- may be used to endorse or promote products derived from this software without
- specific prior written permission. No license is granted to the trademarks of
- the copyright holders even if such marks are included in this software.
-
- THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
- AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
- ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE
- FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
- DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
- SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
- CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
- OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-
-*/
-
-
-/* Text choice questions */
-"CHOICE_OTHER" = "Other"; // Word to use in a muliple choice list that allows 'other' as a choice
-"CHOICE_NONE_OF_THE_ABOVE" = "None of the above"; // Selection choice for none of the above
-"CHOICE_SKIP" = "Prefer not to answer"; // Text for selection when prefer not to answer (skip)
-"CHOICE_NOT_SURE" = "Not sure"; // Text for selection when not sure of the answer
-
-/* Textfield entry */
-"INVALID_REGEX_MESSAGE" = "The text you have entered is invalid."; // Generic message for invalid text entry
-
-/* Passcode Reset */
-"RESET_PASSCODE_TITLE" = "Reset Passcode"; // Prompt to change passcode
-"RESET_PASSCODE_MESSAGE" = "In order to reset your passcode, you'll need to log out of the app completely and log back in using your email and password."; // Description of what happens when passcode is reset
-
-/* Joining items */
-"TWO_ITEM_LIST_FORMAT_%1$@_%2$@" = "%1$@ or %2$@"; // Format of a list with two items (For example, 'Levodopa or Rytary')
-"THREE_ITEM_LIST_FORMAT_%1$@_%2$@_%3$@" = "%1$@, %2$@, or %3$@"; // Format of a list with three items (For example, 'Levodopa, Simet, or Rytary')
-"LIST_FORMAT_DELIMITER"  = ", "; // Delimiter for a list of more than 3 items. Include whitespace as appropriate. (For example, 'Foo, Levodopa, Simet, or Rytary')
-
-/* Time interval */
-"LESS_THAN_%@_AGO" = "Less than %@ ago"; // Less than %@ time interval in the past
-"MORE_THAN_%@_AGO" = "More than %@ ago"; // More than %@ time interval in the past
-"RANGE_%@_AGO" = "%@ ago"; // Range time interval in the past (Ex. '0-30 minutes ago')
-
-/* Progress */
-"CURRENT_STEP_%@_OF_TOTAL_STEPS_%@" = "Step %1$@ of %2$@";  // Text for marking the current steps
-
-/* Default placeholders to request answers in a questionnaire. */
-"PLACEHOLDER_TEXT_OR_NUMBER" = "Tap to answer";
-"NULL_ANSWER" = "Select an answer";
-"PLACEHOLDER_IMAGE_CHOICES" = "Tap to select";
-"PLACEHOLDER_LONG_TEXT" = "Tap to write";
-
-/* Button titles */
-"BUTTON_OK" = "OK";
-"BUTTON_CANCEL" = "Cancel";
-"BUTTON_DONE" = "Done";
-"BUTTON_GET_STARTED" = "Get Started";
-"BUTTON_LEARN_MORE" = "Learn more";
-"BUTTON_NEXT" = "Next";
-"BUTTON_BACK" = "Back";
-"BUTTON_SKIP" = "Prefer not to answer";
-"BUTTON_SKIP_TASK" = "I can't do this now";
-"BUTTON_PAUSE" = "Pause";
-"BUTTON_RESUME" = "Resume";
-
-/* Validation Error Messages */
-"VALIDATION_ERROR_TITLE" = "Invalid Answer";
-"VALIDATION_ERROR_DATE_%@" = "Sorry, the date you entered is not in a recognized format. The expected format is %@";
-"VALIDATION_ERROR_NUMBER" = "Sorry, the number you entered is not in a recognized format.";
-"VALIDATION_ERROR_GENERIC" = "Sorry, the text you have entered is invalid."; // Generic message for invalid text entry
-"VALIDATION_ERROR_MAXLEN_%@" = "The text you entered exceeds the maximum character length of %@ characters.";
-"VALIDATION_ERROR_MIN_DATE_%@" = "The date entered is before %@.";
-"VALIDATION_ERROR_MAX_DATE_%@" = "The date entered is after %@.";
-"VALIDATION_ERROR_MIN_NUMBER_%@" = "The number entered is less than %@.";
-"VALIDATION_ERROR_MAX_NUMBER_%@" = "The number entered is more than %@.";
-
-
